tf/requirements.txt
parliament-functions==0.1.0
Flask==2.0.2
werkzeug==2.0.2
pandas==1.3.3
numpy==1.21.2
azure-storage-blob==12.8.1
psycopg2-binary==2.9.1